Adding .nojekyll file in an attempt to get node_modules working
After running `npm install galleria` and testing connection to the
npm package via index.html, galleria runs successfully in localhost.

However, it doesn't run on Github pages; galleria.js returns
`404 file not found` and the entire node_modules directory is missing.

Using the solution suggested in the following StackOverflow thread:
https://stackoverflow.com/questions/43481269/github-pages-404-on-node-modules-folder

This solution requires the creation of a `gh-pages` branch
and the creation of a blank .nojekyll file in the root of that branch.

I ran `touch .nojekyll` at the root of the Heather directory,
and ran `git checkout -b gh-pages` from the master branch,
and changed the remote repository settings
on Github to point the Github Pages deployment to gh-pages.

Responds to Issue #3.