Added additional HTML for gh-pages testing & a mysterious addition
Added HTML above and below the `script` tag to make sure Github Pages
is taking up stuff pushed to the `gh-pages` branch.

Upon my last commit, a mysterious file called `typescript` was added.

I don't know if this is a result of adding the `.nojekyll` file (see
previous commit for details on that one).

It appears to be a file that just records what I typed in at the command
line. Fairly innocuous, so I'm leaving it for now.

This is starting to feel a bit far around the barn to go to get it working.
If I can't get it working soon, I may default to the method of
installing Galleria.js described in the Beginners' Guide:
http://docs.galleria.io/article/15-beginners-guide

I installed with `npm` instead, because it feels more like a best practice.
But a best practice that I can't get running is...not so interesting.

Responds to Issue #3.