Added styling to gallery
Gallery requires style info in head to determine width & height of gallery.

Added the following in the `<head>` element:

```
<style>
    .galleria{ max-width: 700px; height: 400px; background: #000 }
</style>

```
After adding this, this worked locally. It would be great to have webpack
in this repo so that I could run a localhost. There's an upcoming
Prov.JS meetup on this very topic:

https://www.meetup.com/Prov-JS/events/242770905/

Responds to Issue #6 and Issue #5.